refac algunas cosas

+++ src/controllers/cart.controller.js
@@ -1,27 +1,32 @@
-import { productExist } from '../helper/productExist.helper.js';
-import { stockAvalible } from '../helper/stockAvailable.helper.js';
+import Product from '../models/product.model.js';
+import { stockAvailable } from '../helper/stockAvailable.helper.js';
 import Cart from '../models/cart.model.js';
+import Stock from '../models/stock.model.js';
+import { verifyProductExist } from '../helper/productExist.helper.js';
 
 
 export const insertProductCartController = async(req) =>{
     try{ 
+
         const clientID = req.user.idUser; 
         const quantityCart = req.body.quantityCart;
         const productID = req.params.id;
 
-        const isProductAvailable = await productExist(productID)
-        if(!isProductAvailable){
-            return{
-                data: isProductAvailable
-            }
+        // Validate if the product exists
+        const productData = await Product.productExist(productID)
+        const verifyProduct = verify(productData, productID);
+        if(!verifyProduct){
+            return verifyProduct
         }
-        const theresStock = stockAvalible(quantityCart, productID);
+
+        // Check if the product has stock available
+        const theresStock = stockAvailable(quantityCart, productID);
         if(!theresStock){
             return{
                 data: theresStock
             }
         }
-
+        // Add the product to the cart
         const addToCart = await Cart.createCart(quantityCart, productID, clientID)
         return{
             success: true,
@@ -43,6 +48,20 @@ export const editItemCartController = async(req)=>{
         const {quantityCart, cartID} = req.body;
         const clientID = req.user.idUser;
 
+        // Validate if the product exists
+        const productData = await Product.productExist(cartID);
+        const verifyProduct = verify(productData, cartID);
+        if(!verifyProduct){
+            return verifyProduct
+        }
+        // Validate if the stock is available
+        const stock = await Stock.getStockAvailable(productID)
+        const verifyStock = await stockAvailable(quantityCart, stock)
+        if (!stock.available) {
+            return verifyStock
+        }
+
+        //Edit for quantity product cart
         const editItemCart = await Cart.editItemCart(quantityCart, cartID, clientID)
         return{
             success: true,
@@ -65,6 +84,13 @@ export const deleteProductCartController = async(req) =>{
         const cartID = req.params.id;
         const clientID = req.user.idUser;
 
+        // Validate if the product exists
+        const productData = await Product.productExist(cartID);
+        const verifyProduct = verifyProductExist(productData, cartID);
+         if(!verifyProduct){
+            return verifyProduct
+        }
+        //Delete product cart
         const deleteItemCart = await Cart.deleteProductCart(cartID, clientID);
         return{
             success: true,
